+++ routes/index.js
@@ -1,58 +1,22 @@
 const express = require('express');
-const res = require('express/lib/response');
 const router = express.Router();
-const user = require('../controllers/user');
-const path = require('path');
-
-//multer invoegen
-const multer = require("multer");
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, path.join(__dirname, "../public/uploads") );
-  },
-  filename: function (req, file, callback) {
-    callback(null, file.fieldname + "-" + Date.now() + file.originalname);
-  },
-});
-
-//afbeeldingen kunnen toevoegen
-const upload = multer({ storage: storage});
+const homeRoutes = require('./home');
+const userRoutes = require('./users');
+const loginRoutes = require('./login');
+const registerRoutes = require('./register');
+const goalRoutes = require('./goals');
 
 
-//Hier zet ik alle routes en zo zet ik ze weer naar een controller
-router.get('/', (req,res,) => {
-    res.render('start');
-
-});
-
-//hier import ik de usercontroller en de benodigde functies die ik daarin heb geschreven
-
-router.get('/users', user.fetch);
-
-router.get('/create', user.pass);
-
-router.get('/register', (req,res) => {
-    res.render('register');
-})
+router.use('/', homeRoutes);
+router.use('/users', userRoutes);
+router.use('/register', registerRoutes);
+router.use('/login', loginRoutes);
 
 
 //page not found
 // router.get("*", (req, res) => {
 //     res.send("Not found");
 // });
-
-router.get('/users', user.fetch);
-
-router.post('/filteruser', user.filtereduser);
-
-
-router.post('/login', user.login);
-
-
-router.post('/register', upload.single("picture"), user.register);
-
-// router.post('/register', user.register);
-
 
 
 module.exports = router;
