+++ routes/login.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const authorization = require('../controllers/authorization');
+
+
+router.post('/', authorization.login);
+
+module.exports = router;
