+++ views/register.hbs
@@ -5,7 +5,7 @@
     </a>
     <h1>Create an account</h1>
 
-    <form action="/create" method="get" enctype="multipart/form-data">
+    <form action="/register/create" method="get" enctype="multipart/form-data">
         <label class="mobile-register" for="firstname">First name:</label><br>
         <input type="text"  name="firstname" placeholder="First name" required><br>
